Impl StorePathSecrets
