+++ azure-pipelines.yml
@@ -18,7 +18,7 @@ stages:
       displayName: 'Shutdown'
 
 - stage: Docker_Build_and_Push
-  condition: ne(variables['Build.SourceBranch'], 'refs/heads/master')
+  condition: eq(variables['Build.SourceBranch'], 'refs/heads/master')
   jobs:
   - job: Docker
     pool:
