+++ frame/host/src/lib.rs
@@ -7,9 +7,11 @@ pub mod engine;
 mod config;
 pub mod init_enclave;
 mod ocalls;
+mod store_path_secrets;
 
 pub use error::FrameHostError as Error;
 pub use init_enclave::EnclaveDir;
+pub use store_path_secrets::StorePathSecrets;
 use std::{env, path::PathBuf};
 
 lazy_static! {
